Disable the kaleidoscope example for LLVM 15 (#2)
